dash_clio_dashboard/show_final_user_summary.py
#!/usr/bin/env python3
"""
Final User Consistency Summary

Shows the corrected and consistent user configuration across all three files.
"""

from rich.console import Console
from rich.table import Table
from rich.panel import Panel

console = Console()

def show_final_summary():
    """Show the final corrected user configuration"""
    
    console.print("✅ User Configuration Consistency - Final Status", style="bold green")
    console.print("=" * 80)
    
    # User mapping table
    user_table = Table(show_header=True, header_style="bold magenta")
    user_table.add_column("Email", style="cyan", width=35)
    user_table.add_column("Name", style="yellow", width=20)
    user_table.add_column("CSV Role", style="green", width=15)
    user_table.add_column("YAML Role", style="blue", width=18)
    user_table.add_column("Status", style="white", width=8)
    
    users = [
        ("travis.crawford@cfelab.com", "Travis Crawford", "Administrator", "managing_attorney", "✅"),
        ("lit.attorney@cfelab.com", "Lisa Litigator", "Litigation", "litigation_attorney", "✅"),
        ("lit.assistant@cfelab.com", "Amy Assistant", "Litigation", "litigation_assistant", "✅"),
        ("prelitigation.attorney@cfelab.com", "Paul Prelit", "Prelitigation", "prelitigation_attorney", "✅"),
        ("prelitigation.assistant@cfelab.com", "Nina Assistant", "Prelitigation", "prelitigation_assistant", "✅"),
        ("operations@cfelab.com", "Omar Ops", "Accounts", "operations_manager", "✅"),
        ("intake.department@cfelab.com", "Ivy Intake", "Intake", "intake_specialist", "✅"),
    ]
    
    for email, name, csv_role, yaml_role, status in users:
        user_table.add_row(email, name, csv_role, yaml_role, status)
    
    console.print(user_table)
    
    # Fixed issues summary
    fixes_table = Table.grid(padding=1)
    fixes_table.add_column("Issue", style="cyan", width=30)
    fixes_table.add_column("Before", style="red", width=25)
    fixes_table.add_column("After", style="green", width=25)
    
    fixes_table.add_row(
        "Domain inconsistency",
        "Mixed @cfesolutions.demo",
        "All @cfelab.com"
    )
    
    fixes_table.add_row(
        "Email address mismatches",
        "prelit.attorney vs prelitigation.attorney",
        "All use prelitigation.attorney"
    )
    
    fixes_table.add_row(
        "Missing user references",
        "ivy@ vs intake.department@",
        "All use intake.department@"
    )
    
    fixes_table.add_row(
        "CSV encoding issues",
        "BOM causing parse errors",
        "Clean UTF-8 encoding"
    )
    
    fixes_table.add_row(
        "Mailbox member consistency",
        "Invalid email references",
        "All members exist in user configs"
    )
    
    console.print(Panel(fixes_table, title="🔧 Issues Fixed", border_style="green"))
    
    # Mailbox assignments
    console.print(f"\n📬 Corrected Mailbox Assignments:", style="bold blue")
    
    mailbox_assignments = [
        ("Intake Department", ["intake.department@cfelab.com", "travis.crawford@cfelab.com"]),
        ("Pre-Litigation", ["prelitigation.attorney@cfelab.com", "prelitigation.assistant@cfelab.com", "travis.crawford@cfelab.com"]),
        ("Litigation", ["lit.attorney@cfelab.com", "lit.assistant@cfelab.com", "travis.crawford@cfelab.com"]),
        ("Operations", ["operations@cfelab.com", "travis.crawford@cfelab.com"]),
        ("Firm Info", ["travis.crawford@cfelab.com"]),
    ]
    
    mailbox_table = Table(show_header=True, header_style="bold magenta")
    mailbox_table.add_column("Department", style="cyan", width=20)
    mailbox_table.add_column("Members", style="yellow", width=50)
    
    for dept, members in mailbox_assignments:
        members_str = ", ".join(members)
        mailbox_table.add_row(dept, members_str)
    
    console.print(mailbox_table)
    
    # Validation results
    console.print(f"\n🎯 Validation Results:", style="bold green")
    
    validation_results = [
        "✅ All 7 users consistent across CSV and YAML",
        "✅ Single domain (@cfelab.com) used throughout",
        "✅ All mailbox members reference valid users",
        "✅ Role mappings align between subscription types and YAML roles",
        "✅ No duplicate or conflicting email addresses",
        "✅ Clean CSV encoding without BOM issues",
        "✅ Proper departmental groupings in mailbox configuration"
    ]
    
    for result in validation_results:
        console.print(f"   {result}")
    
    console.print(f"\n🚀 Configuration is now ready for production deployment!", style="bold green")

if __name__ == "__main__":
    show_final_summary()