requirements-cfe.txt
# CFE Solutions Integration Requirements
# Additional dependencies for standalone services integration

# Core Dashboard Requirements
dash==2.14.1
dash-bootstrap-components==1.5.0
dash-mantine-components==0.12.1
plotly==5.17.0
pandas==2.1.3
requests==2.31.0
redis==5.0.1

# gRPC and Protocol Buffers
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-health-checking==1.60.0
grpcio-reflection==1.60.0
protobuf==4.25.1

# Async gRPC support
grpcio-status==1.60.0
googleapis-common-protos==1.61.0

# Enhanced service discovery
consul-python==1.1.0
python-consul==1.1.0

# Configuration management
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Authentication and security
PyJWT==2.8.0
cryptography==41.0.8
bcrypt==4.1.2

# Enhanced Neo4j integration
neo4j==5.15.0
py2neo==2021.2.3

# ChromaDB client
chromadb==0.4.18

# Enhanced HTTP client for service communication
httpx==0.25.2
aiohttp==3.9.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing (dev profile)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Performance monitoring
psutil==5.9.6
memory-profiler==0.61.0